Projeto_Integrador/projeto_integrador_7/Flask_MQTT/requirements.txt
Flask
Flask-Admin
Flask-Cors
Flask-JWT
Flask-Login
Flask-Migrate
Flask-MQTT
Flask-RESTful
Flask-SQLAlchemy
Flask-WTF
Jinja2
passlib
py3-validate-email
psycopg2-binary
python-dotenv
requests
wtforms[email]
